moved deserialization tests to separate test
